backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers